+++ Scripts/clear-telemetry-qgc.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+#Author: Jesimar da Silva Arantes
+#Date: 22/02/2018
+#Last Update: 22/02/2018
+
+echo "====================Clear Files Telemetry QGC====================="
+
+rm /home/jesimar/Documentos/QGroundControl/Telemetry/*.tlog
+
+echo "==============================Done================================"