+++ Scripts/info-hd.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-#Author: Jesimar da Silva Arantes
-#Date: 12/01/2018
-#Last Update: 12/01/2018
-
-echo "==================Informations HD=================="
-df -h