+++ Scripts/list-info-hd.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+#Author: Jesimar da Silva Arantes
+#Date: 12/01/2018
+#Last Update: 23/02/2018
+#Description: Script that displays HD usage information.
+#Descrição: Script que exibe informações de uso do HD.
+
+echo "==================HD Informations=================="
+
+df -h
+
+echo "=======================Done========================"