+++ Scripts/my-ip.sh
@@ -2,6 +2,8 @@
 #Author: Jesimar da Silva Arantes
 #Date: 19/10/2017
 #Last Update: 21/02/2018
+#Description: Script that filters information and displays the IP of the computer.
+#Descrição: Script que filtra informações e exibe apenas o IP do computador.
 
 echo 'my ip:'